refactor(output): split formatting into output/{human,json,templates} [roadmap:v0.7.4]
Split the monolithic outputs.py into the rac.output package: human-readable renderers in output/human.py, JSON in output/json.py, Markdown templates in output/templates.py, with the two shared improve-guidance fallbacks in output/_shared.py. output/__init__ re-exports the full render_* surface, and a shim at rac/outputs.py forwards to it so existing imports keep resolving. JSON contracts (ADR-007) are unchanged.
